+++ src/GTP.h
@@ -90,7 +90,7 @@ extern size_t cfg_max_tree_size;
 extern int cfg_max_cache_ratio_percent;
 extern TimeManagement::enabled_t cfg_timemanage;
 extern int cfg_lagbuffer_cs;
-extern int cfg_resignpct;
+extern float cfg_resignpct;
 extern int cfg_noise;
 extern int cfg_random_cnt;
 extern int cfg_random_min_visits;
